lots of random stuff. Not working currently.

+++ movievultures/src/main/java/movievultures/model/dao/UserDao.java
@@ -5,7 +5,7 @@ import java.util.List;
 import movievultures.model.User;
 
 public interface UserDao {
-	User getUser(Long id);
+	User getUser(int id);
 	User getUserByUsername(String username);
 	List<User> getUsersByUsername(String username);
 	User saveUser(User user);
